feat(link-purpose-template-agevolata): [QA-14284] moved some methods to static helper
Fixed a few typos and removed obsolete comments and imports
